ARC-744 Moved start and stop functions from worker.ts to a separate file
